레시피 수정시 이미지 변경이 없으면, 기존 이미지 유지하도록 수정

+++ src/main/java/team/rescue/recipe/service/RecipeService.java
@@ -228,7 +228,8 @@ public class RecipeService {
 
 		// 대표 이미지 수정
 		String recipeImageFilePath = recipe.getRecipeImageUrl();
-		if (recipeImage != null && !recipeImage.isEmpty()) {
+		if (recipeImage != null && !recipeImage.isEmpty() &&
+				!Objects.equals(recipeImage.getContentType(), "String")) {
 			// 레시피 대표 이미지 업데이트
 			fileService.deleteImages(recipeImageFilePath);
 			recipeImageFilePath = fileService.uploadImageToS3(recipeImage);
